feat: header changed and pages access added
