--- app/_components/Playground/index.ts
@@ -1,1 +1,0 @@
-export { Playground } from "./Playground";
