+++ app/_components/index.ts
@@ -7,3 +7,4 @@ export { InfoPanel, TxInfo, UtxoInfo } from "./InfoPanel";
 export { CustomInput as Input } from "./Input";
 export { Playground } from "./Playground";
 export { Line, Transaction, Utxo } from "./Transaction";
+export { TxInput } from "./TxInput";
